+++ examples/example-flutter/android/app/src/main/kotlin/com/example/expression_ui_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.expression_ui_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()