+++ examples/example-flutter/android/app/src/main/kotlin/com/example/rive_counter/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.rive_counter
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity()