+++ examples/example-flutter/lib/app/app.router.dart
@@ -1,0 +1,289 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// StackedNavigatorGenerator
+// **************************************************************************
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:expression_ui_example/ui/views/composable/composable_view.dart'
+    as _i7;
+import 'package:expression_ui_example/ui/views/cross_framework/cross_framework_view.dart'
+    as _i4;
+import 'package:expression_ui_example/ui/views/developer_friendly/developer_friendly_view.dart'
+    as _i5;
+import 'package:expression_ui_example/ui/views/home/home_view.dart' as _i2;
+import 'package:expression_ui_example/ui/views/pixel_perfection/pixel_perfection_view.dart'
+    as _i6;
+import 'package:expression_ui_example/ui/views/startup/startup_view.dart'
+    as _i3;
+import 'package:flutter/material.dart' as _i8;
+import 'package:flutter/material.dart';
+import 'package:stacked/stacked.dart' as _i1;
+import 'package:stacked_services/stacked_services.dart' as _i9;
+
+class Routes {
+  static const homeView = '/home-view';
+
+  static const startupView = '/startup-view';
+
+  static const crossFrameworkView = '/cross-framework-view';
+
+  static const developerFriendlyView = '/developer-friendly-view';
+
+  static const pixelPerfectionView = '/pixel-perfection-view';
+
+  static const composableView = '/composable-view';
+
+  static const all = <String>{
+    homeView,
+    startupView,
+    crossFrameworkView,
+    developerFriendlyView,
+    pixelPerfectionView,
+    composableView,
+  };
+}
+
+class StackedRouter extends _i1.RouterBase {
+  final _routes = <_i1.RouteDef>[
+    _i1.RouteDef(
+      Routes.homeView,
+      page: _i2.HomeView,
+    ),
+    _i1.RouteDef(
+      Routes.startupView,
+      page: _i3.StartupView,
+    ),
+    _i1.RouteDef(
+      Routes.crossFrameworkView,
+      page: _i4.CrossFrameworkView,
+    ),
+    _i1.RouteDef(
+      Routes.developerFriendlyView,
+      page: _i5.DeveloperFriendlyView,
+    ),
+    _i1.RouteDef(
+      Routes.pixelPerfectionView,
+      page: _i6.PixelPerfectionView,
+    ),
+    _i1.RouteDef(
+      Routes.composableView,
+      page: _i7.ComposableView,
+    ),
+  ];
+
+  final _pagesMap = <Type, _i1.StackedRouteFactory>{
+    _i2.HomeView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i2.HomeView(),
+        settings: data,
+      );
+    },
+    _i3.StartupView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i3.StartupView(),
+        settings: data,
+      );
+    },
+    _i4.CrossFrameworkView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i4.CrossFrameworkView(),
+        settings: data,
+      );
+    },
+    _i5.DeveloperFriendlyView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i5.DeveloperFriendlyView(),
+        settings: data,
+      );
+    },
+    _i6.PixelPerfectionView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i6.PixelPerfectionView(),
+        settings: data,
+      );
+    },
+    _i7.ComposableView: (data) {
+      return _i8.MaterialPageRoute<dynamic>(
+        builder: (context) => const _i7.ComposableView(),
+        settings: data,
+      );
+    },
+  };
+
+  @override
+  List<_i1.RouteDef> get routes => _routes;
+
+  @override
+  Map<Type, _i1.StackedRouteFactory> get pagesMap => _pagesMap;
+}
+
+extension NavigatorStateExtension on _i9.NavigationService {
+  Future<dynamic> navigateToHomeView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.homeView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> navigateToStartupView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.startupView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> navigateToCrossFrameworkView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.crossFrameworkView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> navigateToDeveloperFriendlyView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.developerFriendlyView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> navigateToPixelPerfectionView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.pixelPerfectionView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> navigateToComposableView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return navigateTo<dynamic>(Routes.composableView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithHomeView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.homeView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithStartupView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.startupView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithCrossFrameworkView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.crossFrameworkView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithDeveloperFriendlyView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.developerFriendlyView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithPixelPerfectionView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.pixelPerfectionView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+
+  Future<dynamic> replaceWithComposableView([
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  ]) async {
+    return replaceWith<dynamic>(Routes.composableView,
+        id: routerId,
+        preventDuplicates: preventDuplicates,
+        parameters: parameters,
+        transition: transition);
+  }
+}