+++ examples/example-flutter/lib/main.dart
@@ -1,13 +1,16 @@
-import 'package:device_preview/device_preview.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:rive_counter/tiles/list_tiles_view.dart';
+import 'package:expression_ui_example/app/app.bottomsheets.dart';
+import 'package:expression_ui_example/app/app.dialogs.dart';
+import 'package:expression_ui_example/app/app.locator.dart';
+import 'package:expression_ui_example/app/app.router.dart';
+import 'package:stacked_services/stacked_services.dart';
 
-void main() {
-  runApp(DevicePreview(
-    enabled: !kReleaseMode,
-    builder: (context) => const MainApp()
-  ));
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await setupLocator();
+  setupDialogUi();
+  setupBottomSheetUi();
+  runApp(const MainApp());
 }
 
 class MainApp extends StatelessWidget {
@@ -15,9 +18,13 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return  MaterialApp(
-       home: ListTilesView(),
-      //home: ResponsiveView(),
+    return MaterialApp(
+      initialRoute: Routes.startupView,
+      onGenerateRoute: StackedRouter().onGenerateRoute,
+      navigatorKey: StackedService.navigatorKey,
+      navigatorObservers: [
+        StackedService.routeObserver,
+      ],
     );
   }
 }