+++ examples/example-flutter/lib/ui/views/startup/startup_view.dart
@@ -1,0 +1,29 @@
+import 'package:expression_ui_example/expression_ui/expression_view.dart';
+import 'package:flutter/material.dart';
+import 'package:stacked/stacked.dart';
+
+import 'startup_viewmodel.dart';
+
+class StartupView extends StackedView<StartupViewModel> {
+  const StartupView({Key? key}) : super(key: key);
+
+  @override
+  Widget builder(
+    BuildContext context,
+    StartupViewModel viewModel,
+    Widget? child,
+  ) {
+    return Scaffold(
+      body: ExpressionView(
+        artboardName: 'intro-view',
+        onEvent: (_) => viewModel.navigateToNextView(),
+      ),
+    );
+  }
+
+  @override
+  StartupViewModel viewModelBuilder(
+    BuildContext context,
+  ) =>
+      StartupViewModel();
+}