+++ examples/example-flutter/test/viewmodels/home_viewmodel_test.dart
@@ -1,0 +1,40 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/mockito.dart';
+import 'package:expression_ui_example/app/app.bottomsheets.dart';
+import 'package:expression_ui_example/app/app.locator.dart';
+import 'package:expression_ui_example/ui/common/app_strings.dart';
+import 'package:expression_ui_example/ui/views/home/home_viewmodel.dart';
+
+import '../helpers/test_helpers.dart';
+
+void main() {
+  HomeViewModel getModel() => HomeViewModel();
+
+  group('HomeViewmodelTest -', () {
+    setUp(() => registerServices());
+    tearDown(() => locator.reset());
+
+    group('incrementCounter -', () {
+      test('When called once should return  Counter is: 1', () {
+        final model = getModel();
+        model.incrementCounter();
+        expect(model.counterLabel, 'Counter is: 1');
+      });
+    });
+
+    group('showBottomSheet -', () {
+      test('When called, should show custom bottom sheet using notice variant',
+          () {
+        final bottomSheetService = getAndRegisterBottomSheetService();
+
+        final model = getModel();
+        model.showBottomSheet();
+        verify(bottomSheetService.showCustomSheet(
+          variant: BottomSheetType.notice,
+          title: ksHomeBottomSheetTitle,
+          description: ksHomeBottomSheetDescription,
+        ));
+      });
+    });
+  });
+}