+++ examples/example-react-native/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  root: true,
+  extends: '@react-native',
+  parserOptions: {
+    requireConfigFile: false,
+  },
+};