+++ packages/expression-ui-flutter/lib/expression_ui.dart
@@ -1,6 +1,5 @@
 library expression_ui;
 
 export 'src/expression_widgets/base_expression_view.dart';
-export 'src/models/tap_event.dart';
-export 'src/expression_widgets/list_tile/expression_ui_to_json_tile.dart';
-export 'src/expression_widgets/list_tile/tile_widget.dart';
+export 'src/models/state_event.dart';
+export 'src/utils/state_controller.dart';