+++ packages/expression-ui-flutter/lib/src/expression_widgets/base_expression_view.dart
@@ -1,6 +1,8 @@
 import 'package:expression_ui/src/extensions/rive_extensions.dart';
-import 'package:expression_ui/src/models/tap_event.dart';
+import 'package:expression_ui/src/models/state_event.dart';
+import 'package:expression_ui/src/utils/state_controller.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:rive/rive.dart';
 
 class BaseExpressionView extends StatefulWidget {
@@ -13,43 +15,53 @@ class BaseExpressionView extends StatefulWidget {
   /// Fired when an event pointer down happens in the rive file
   ///
   /// Events has the be triggered in the rive file
-  final Function(TapEvent) onTap;
+  final Function(StateEvent) onEvent;
 
   /// Set to true if the entire view needs to be scrollable
   ///
   /// defaults to false
+  /// Alias for [useArtboardSize]
   final bool scrollable;
 
   /// The name of the state machine to load
   final String? stateMachine;
 
+  /// The map of dynamic text updates to perform
+  final Map<String, String> textValues;
+
+  /// Controller to manage the state of [artboardName]
+  final StateController? controller;
+
+  final BoxFit fit;
+
+  /// Internally sizes the wdget to match the artboard dimensions defined
+  /// on the canvas
+  final bool useArtboardSize;
+
   const BaseExpressionView({
     super.key,
     required this.filePath,
     required this.artboardName,
-    required this.onTap,
+    required this.onEvent,
+    this.useArtboardSize = false,
     this.scrollable = false,
     this.stateMachine,
+    this.textValues = const {},
+    this.controller,
+    this.fit = BoxFit.cover,
   });
-
-  @override
-  State<BaseExpressionView> createState() => BaseExpressionViewState();
-}
-
-class BaseExpressionViewState extends State<BaseExpressionView> {
-  late final Artboard _artboard;
 
   @override
   Widget build(BuildContext context) {
     final riveView = RiveAnimation.asset(
-      widget.filePath,
-      artboard: widget.artboardName,
-      useArtboardSize: widget.scrollable,
-      fit: BoxFit.cover,
+      filePath,
+      artboard: artboardName,
+      useArtboardSize: useArtboardSize || scrollable,
+      fit: fit,
       onInit: onRiveFileInitialized,
     );
 
-    if (!widget.scrollable) {
+    if (!scrollable) {
       return riveView;
     }
 
@@ -64,22 +76,29 @@ class BaseExpressionViewState extends State<BaseExpressionView> {
   }
 
   void onRiveFileInitialized(Artboard artboard) {
-    _artboard = artboard;
-    final controller = StateMachineController.fromArtboard(
+    final stateMachineController = StateMachineController.fromArtboard(
       artboard,
-      widget.stateMachine ?? 'State Machine 1',
+      stateMachine ?? 'State Machine 1',
     );
 
-    if (controller != null) {
-      controller.addEventListener((event) {
-        widget.onTap.call(TapEvent(eventName: event.name));
+    for (var entry in textValues.keys) {
+      final textRun = artboard.textRun(entry);
+
+      if (textRun != null) {
+        textRun.text = textValues[entry]!;
+      }
+    }
+
+    if (stateMachineController != null) {
+      stateMachineController.addEventListener((event) {
+        SchedulerBinding.instance.addPostFrameCallback((_) {
+          onEvent.call(StateEvent(name: event.name));
+        });
       });
 
-      artboard.addController(controller);
-    }
-  }
+      controller?.setInputs(stateMachineController.inputs);
 
-  void editTextValueRun(String key, String value) {
-    _artboard.textRun(key)!.text = value;
+      artboard.addController(stateMachineController);
+    }
   }
 }