+++ packages/expression-ui-flutter/lib/src/utils/state_controller.dart
@@ -1,0 +1,18 @@
+import 'package:rive/rive.dart';
+
+class StateController {
+  Map<String, SMIInput> inputs = {};
+
+  void setInputs(Iterable<SMIInput> inputsFromArtboard) {
+    for (var input in inputsFromArtboard) {
+      inputs[input.name] = input;
+    }
+  }
+
+  void fireTrigger({required String name}) {
+    final input = inputs[name];
+    if (input != null && input is SMITrigger) {
+      input.fire();
+    }
+  }
+}