+++ packages/expression-ui-flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+-feat: Adds `StateController` that allows you to trigger inputs from your Flutter app
+-feat: Adds `textValues` map, that sets the value of a text run on start
+
 ## 0.1.0
 
 - feat: Adds `BaseExpressionView` with scrollable functionality and basic event tap listening